[feat](utils) move GitUtils
